compressed columns, adapting to cohomology + multi fields

+++ src/Persistence_matrix/include/gudhi/column_types/row_access.h
@@ -24,6 +24,7 @@ template<class Row_container_type, class Cell_type, bool isIntrusive, bool hasRe
 class Row_access
 {
 public:
+	Row_access();
 	Row_access(index columnIndex, Row_container_type& rows);
 	Row_access(Row_access&& other) noexcept;
 
@@ -45,13 +46,10 @@ protected:
 	index columnIndex_;
 	Row_container_type* rows_;					//be carefull to not destroy container before columns
 	static constexpr bool isActive_ = true;
-
-private:
-	Row_access();	//for more explicit error message when compiling
 };
 
 template<class Row_container_type, class Cell_type, bool isIntrusive, bool hasRemovableColumns>
-inline Row_access<Row_container_type,Cell_type,isIntrusive,hasRemovableColumns>::Row_access()
+inline Row_access<Row_container_type,Cell_type,isIntrusive,hasRemovableColumns>::Row_access() : rows_(nullptr)
 {}
 
 template<class Row_container_type, class Cell_type, bool isIntrusive, bool hasRemovableColumns>
@@ -68,16 +66,29 @@ inline Row_access<Row_container_type,Cell_type,isIntrusive,hasRemovableColumns>:
 template<class Row_container_type, class Cell_type, bool isIntrusive, bool hasRemovableColumns>
 inline void Row_access<Row_container_type,Cell_type,isIntrusive,hasRemovableColumns>::insert_cell(index rowIndex, Cell_type *cell)
 {
+	if (rows_ == nullptr) return;
+
 	if constexpr (hasRemovableColumns)
 		rows_->try_emplace(rowIndex);
 	else {
 		if (rows_->size() < rowIndex + 1)
 			rows_->resize(rowIndex + 1);
+//		while (rows_->size() < rowIndex + 1) {
+//			rows_->emplace_back();
+//		}
 	}
 	if constexpr (isIntrusive){
-		rows_->at(rowIndex).push_back(*cell);
+		if constexpr (hasRemovableColumns)
+			rows_->at(rowIndex).push_back(*cell);
+		else {
+			rows_->operator[](rowIndex).push_back(*cell);
+		}
 	} else {
-		rows_->at(rowIndex).insert(*cell);
+		if constexpr (hasRemovableColumns)
+			rows_->at(rowIndex).insert(*cell);
+		else {
+			rows_->operator[](rowIndex).insert(*cell);
+		}
 	}
 }
 
@@ -86,18 +97,23 @@ inline void Row_access<Row_container_type,Cell_type,isIntrusive,hasRemovableColu
 {
 	static_assert(!isIntrusive, "Cannot insert const cell in intrusive container.");
 
+	if (rows_ == nullptr) return;
+
 	if constexpr (hasRemovableColumns)
 		rows_->try_emplace(rowIndex);
 	else {
 		if (rows_->size() < rowIndex + 1)
 			rows_->resize(rowIndex + 1);
 	}
+
 	rows_->at(rowIndex).insert(cell);
 }
 
 template<class Row_container_type, class Cell_type, bool isIntrusive, bool hasRemovableColumns>
 inline void Row_access<Row_container_type,Cell_type,isIntrusive,hasRemovableColumns>::unlink(Cell_type *cell)
 {
+	if (rows_ == nullptr) return;
+
 	if constexpr (isIntrusive){
 		cell->base_hook_matrix_row::unlink();
 	} else {
@@ -110,6 +126,8 @@ inline void Row_access<Row_container_type,Cell_type,isIntrusive,hasRemovableColu
 {
 	static_assert(!isIntrusive, "Cannot unlink const cell from intrusive container.");
 
+	if (rows_ == nullptr) return;
+
 	rows_->at(cell.get_row_index()).erase(cell);
 }
 
@@ -117,8 +135,14 @@ template<class Row_container_type, class Cell_type, bool isIntrusive, bool hasRe
 inline void Row_access<Row_container_type,Cell_type,isIntrusive,hasRemovableColumns>::update_cell(const Cell_type &cell)
 {
 	if constexpr (!isIntrusive){
-		rows_->at(cell.get_row_index()).erase(cell);
-		rows_->at(cell.get_row_index()).insert(cell);
+		if (rows_ == nullptr) return;
+
+		auto& row = rows_->at(cell.get_row_index());
+		auto it = row.find(cell);
+		it = row.erase(it);
+		row.insert(it, cell);
+//		rows_->at(cell.get_row_index()).erase(cell);
+//		rows_->at(cell.get_row_index()).insert(cell);
 	}
 }
