+++ src/Persistence_matrix/include/gudhi/column_types/set_column.h
@@ -81,6 +81,11 @@ public:
 		return column;
 	}
 
+	//this = v * this + column
+	Set_column& multiply_and_add(const Field_element_type& v, const Set_column& column);
+	//this = this + column * v
+	Set_column& multiply_and_add(const Set_column& column, const Field_element_type& v);
+
 	friend bool operator==(const Set_column& c1, const Set_column& c2){
 		auto it1 = c1.column_.begin();
 		auto it2 = c2.column_.begin();
@@ -119,13 +124,14 @@ protected:
 	Column_type column_;
 
 	void _delete_cell(iterator& it);
-	void _insert_cell(const Field_element_type& value, index rowIndex);
+	void _insert_cell(const Field_element_type& value, index rowIndex, const iterator &position);
+	void _clear();
 };
 
 template<class Field_element_type, class Cell_type, class Row_access_option>
 inline Set_column<Field_element_type,Cell_type,Row_access_option>::Set_column() : dim_(0)
 {
-	static_assert(!Row_access_option::isActive_, "When row access option enabled, a row container has to be provided.");
+//	static_assert(!Row_access_option::isActive_, "When row access option enabled, a row container has to be provided.");
 }
 
 template<class Field_element_type, class Cell_type, class Row_access_option>
@@ -133,10 +139,10 @@ template<class Container_type>
 inline Set_column<Field_element_type,Cell_type,Row_access_option>::Set_column(const Container_type &nonZeroRowIndices)
 	: dim_(nonZeroRowIndices.size() == 0 ? 0 : nonZeroRowIndices.size() - 1)
 {
-	static_assert(!Row_access_option::isActive_, "When row access option enabled, a row container has to be provided.");
+//	static_assert(!Row_access_option::isActive_, "When row access option enabled, a row container has to be provided.");
 
-	for (const std::pair<index,Field_element_type>& p : nonZeroRowIndices){
-		_insert_cell(p.second, p.first);
+	for (const auto& p : nonZeroRowIndices){
+		_insert_cell(p.second, p.first, column_.end());
 	}
 }
 
@@ -145,10 +151,10 @@ template<class Container_type>
 inline Set_column<Field_element_type,Cell_type,Row_access_option>::Set_column(const Container_type &nonZeroRowIndices, dimension_type dimension)
 	: dim_(dimension)
 {
-	static_assert(!Row_access_option::isActive_, "When row access option enabled, a row container has to be provided.");
+//	static_assert(!Row_access_option::isActive_, "When row access option enabled, a row container has to be provided.");
 
-	for (const std::pair<index,Field_element_type>& p : nonZeroRowIndices){
-		_insert_cell(p.second, p.first);
+	for (const auto& p : nonZeroRowIndices){
+		_insert_cell(p.second, p.first, column_.end());
 	}
 }
 
@@ -165,8 +171,8 @@ inline Set_column<Field_element_type,Cell_type,Row_access_option>::Set_column(
 		index columnIndex, const Container_type &nonZeroRowIndices, Row_container_type &rowContainer)
 	: Row_access_option(columnIndex, rowContainer), dim_(nonZeroRowIndices.size() == 0 ? 0 : nonZeroRowIndices.size() - 1)
 {
-	for (const std::pair<index,Field_element_type>& p : nonZeroRowIndices){
-		_insert_cell(p.second, p.first);
+	for (const auto& p : nonZeroRowIndices){
+		_insert_cell(p.second, p.first, column_.end());
 	}
 }
 
@@ -176,8 +182,8 @@ inline Set_column<Field_element_type,Cell_type,Row_access_option>::Set_column(
 		index columnIndex, const Container_type &nonZeroRowIndices, dimension_type dimension, Row_container_type &rowContainer)
 	: Row_access_option(columnIndex, rowContainer), dim_(dimension)
 {
-	for (const std::pair<index,Field_element_type>& p : nonZeroRowIndices){
-		_insert_cell(p.second, p.first);
+	for (const auto& p : nonZeroRowIndices){
+		_insert_cell(p.second, p.first, column_.end());
 	}
 }
 
@@ -197,7 +203,7 @@ inline Set_column<Field_element_type,Cell_type,Row_access_option>::Set_column(
 	  dim_(column.dim_)
 {
 	for (const Cell& cell : column.column_){
-		_insert_cell(cell.get_element(), cell.get_row_index());
+		_insert_cell(cell.get_element(), cell.get_row_index(), column_.end());
 	}
 }
 
@@ -342,9 +348,10 @@ inline Set_column<Field_element_type,Cell_type,Row_access_option> &Set_column<Fi
 			Field_element_type coef = c->get_element();
 			coef += v.get_element();
 			_delete_cell(c);
-			if (coef != 0u) _insert_cell(coef, r);
+			if (coef != Field_element_type::get_additive_identity())
+				_insert_cell(coef, r, column_.end());
 		} else
-			_insert_cell(v.get_element(), v.get_row_index());
+			_insert_cell(v.get_element(), v.get_row_index(), column_.end());
 	}
 
 	return *this;
@@ -353,14 +360,17 @@ inline Set_column<Field_element_type,Cell_type,Row_access_option> &Set_column<Fi
 template<class Field_element_type, class Cell_type, class Row_access_option>
 inline Set_column<Field_element_type,Cell_type,Row_access_option> &Set_column<Field_element_type,Cell_type,Row_access_option>::operator*=(unsigned int v)
 {
-	v %= Field_element_type::get_characteristic();
+//	v %= Field_element_type::get_characteristic();
+	Field_element_type val(v);
+
+	if (val == Field_element_type::get_multiplicative_identity()) return *this;
 
 	if constexpr (Row_access_option::isActive_){
 		for (const Cell& cell : column_)
 			Row_access_option::unlink(cell);
 	}
 
-	if (v == 0) {
+	if (val == 0u) {
 		column_.clear();
 		return *this;
 	}
@@ -369,7 +379,7 @@ inline Set_column<Field_element_type,Cell_type,Row_access_option> &Set_column<Fi
 
 	for (const Cell& cell : column_){
 		Cell newCell(cell);
-		newCell.get_element() *= v;
+		newCell.get_element() *= val;
 
 		if constexpr (Row_access_option::isActive_){
 			auto it = newColumn.insert(newCell);
@@ -380,6 +390,77 @@ inline Set_column<Field_element_type,Cell_type,Row_access_option> &Set_column<Fi
 	}
 
 	column_.swap(newColumn);
+
+	return *this;
+}
+
+template<class Field_element_type, class Cell_type, class Row_access_option>
+inline Set_column<Field_element_type,Cell_type,Row_access_option> &
+Set_column<Field_element_type,Cell_type,Row_access_option>::multiply_and_add(const Field_element_type& val, const Set_column& column)
+{
+	if (val == 0u) {
+		_clear();
+	}
+
+	const Column_type& sc = column.column_;
+
+	auto itTarget = column_.begin();
+	auto itSource = sc.begin();
+	while (itTarget != column_.end() && itSource != sc.end())
+	{
+		if (itTarget->get_row_index() < itSource->get_row_index()) {
+			itTarget->get_element() *= val;
+			++itTarget;
+		} else if (itTarget->get_row_index() > itSource->get_row_index()) {
+			_insert_cell(itSource->get_element(), itSource->get_row_index(), itTarget);
+			++itSource;
+		} else {
+			itTarget->get_element() *= val;
+			itTarget->get_element() += itSource->get_element();
+			if (itTarget->get_element() == Field_element_type::get_additive_identity()){
+				_delete_cell(itTarget);
+			} else {
+				if constexpr (Row_access_option::isActive_)
+						Row_access_option::update_cell(*itTarget);
+				++itTarget;
+			}
+			++itSource;
+		}
+	}
+
+	while (itTarget != column_.end()){
+		itTarget->get_element() *= val;
+		++itTarget;
+	}
+
+	while (itSource != sc.end()) {
+		_insert_cell(itSource->get_element(), itSource->get_row_index(), column_.end());
+		++itSource;
+	}
+
+	return *this;
+}
+
+template<class Field_element_type, class Cell_type, class Row_access_option>
+inline Set_column<Field_element_type,Cell_type,Row_access_option> &
+Set_column<Field_element_type,Cell_type,Row_access_option>::multiply_and_add(const Set_column& column, const Field_element_type& val)
+{
+	if (val == 0u) {
+		return *this;
+	}
+
+	for (const Cell& v : column.column_){
+		auto c = column_.find(v);
+		if (c != column_.end()){
+			index r = c->get_row_index();
+			Field_element_type coef(c->get_element());
+			coef += (v.get_element() * val);
+			_delete_cell(c);
+			if (coef != Field_element_type::get_additive_identity())
+				_insert_cell(coef, r, column_.end());
+		} else
+			_insert_cell(v.get_element() * val, v.get_row_index(), column_.end());
+	}
 
 	return *this;
 }
@@ -404,14 +485,24 @@ inline void Set_column<Field_element_type,Cell_type,Row_access_option>::_delete_
 
 template<class Field_element_type, class Cell_type, class Row_access_option>
 inline void Set_column<Field_element_type,Cell_type,Row_access_option>::_insert_cell(
-		const Field_element_type &value, index rowIndex)
+		const Field_element_type &value, index rowIndex, const iterator &position)
 {
 	if constexpr (Row_access_option::isActive_){
-		auto it = column_.emplace_hint(column_.end(), value, Row_access_option::columnIndex_, rowIndex);
+		auto it = column_.emplace_hint(position, value, Row_access_option::columnIndex_, rowIndex);
 		Row_access_option::insert_cell(rowIndex, *it);
 	} else {
-		column_.emplace_hint(column_.end(), value, rowIndex);
+		column_.emplace_hint(position, value, rowIndex);
 	}
+}
+
+template<class Field_element_type, class Cell_type, class Row_access_option>
+inline void Set_column<Field_element_type,Cell_type,Row_access_option>::_clear()
+{
+	if constexpr (Row_access_option::isActive_){
+		for (const Cell& cell : column_)
+			Row_access_option::unlink(cell);
+	}
+	column_.clear();
 }
 
 } //namespace persistence_matrix
@@ -424,7 +515,7 @@ struct std::hash<Gudhi::persistence_matrix::Set_column<Field_element_type,Cell_t
 	{
 		std::size_t seed = 0;
 		for (auto& cell : column){
-			seed ^= std::hash<unsigned int>()(cell.get_row_index() * cell.get_element()) + 0x9e3779b9 + (seed << 6) + (seed >> 2);
+			seed ^= std::hash<unsigned int>()(cell.get_row_index() * static_cast<unsigned int>(cell.get_element())) + 0x9e3779b9 + (seed << 6) + (seed >> 2);
 		}
 		return seed;
 	}
